image-to-pdf/pyproject.toml
[tool.poetry]
name = "@scripts/image-to-pdf"
version = "0.1.0"
description = ""
authors = ["Eugene Moshchyn <eugene.moshchyn@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
